[mediaType] [spanCount] [maxSelectable]

+++ app/src/main/java/com/neurotech/photobrowser/loader/AudioLoader.java
@@ -44,7 +44,7 @@ public class AudioLoader {
         audioBean.setPath(cursor.getString(cursor.getColumnIndex(MediaStore.Audio.AudioColumns.DATA)));
         audioBean.setSize(cursor.getLong(cursor.getColumnIndex(MediaStore.Audio.AudioColumns.SIZE)));
         audioBean.setMimeType(cursor.getString(cursor.getColumnIndex(MediaStore.Audio.AudioColumns.MIME_TYPE)));
-        audioBean.setDuration(cursor.getLong(cursor.getColumnIndex(MediaStore.Audio.AudioColumns.DURATION)));
+        audioBean.setDuration(cursor.getInt(cursor.getColumnIndex(MediaStore.Audio.AudioColumns.DURATION)));
         audioBean.setAlbumId(cursor.getInt(cursor.getColumnIndex(MediaStore.Audio.AudioColumns.ALBUM_ID)));
 
         audioBean.setMediaMimeType(MimeType.AUDIO);
